enhanced doc, first class exported via getClass()
The doc has been cleaned up quite a bit.

Changed the way the first class is exported.  Used to *become*
the export object, but of course that's a little sketchy,
since not every CommonJS runtime supports that.  Instead, a single
function will be exported in the module when the first class
definition is encountered.  The function exported is named 'getClass',
and it will return that first defined class.  The 'requireClass'
directive has been changed to take this into account.